+++ packages/graphiql-toolkit/src/index.ts
@@ -1,2 +1,3 @@
+export * from './async-helpers';
 export * from './create-fetcher';
 // TODO: move the most useful utilities from graphiql to here
