+++ packages/graphiql/src/components/GraphiQL.tsx
@@ -61,12 +61,16 @@ import setValue from 'set-value';
 import type {
   Fetcher,
   FetcherResult,
-  FetcherReturnType,
   FetcherOpts,
   SyncFetcherResult,
-  Observable,
   Unsubscribable,
   FetcherResultPayload,
+} from '@graphiql/toolkit';
+import {
+  fetcherReturnToPromise,
+  isPromise,
+  isObservable,
+  isAsyncIterable,
 } from '@graphiql/toolkit';
 import HistoryStore from '../utility/HistoryStore';
 
@@ -2218,88 +2222,6 @@ const defaultQuery = `# Welcome to GraphiQL
 #
 
 `;
-
-// Duck-type promise detection.
-function isPromise<T>(value: Promise<T> | any): value is Promise<T> {
-  return typeof value === 'object' && typeof value.then === 'function';
-}
-
-// Duck-type Observable.take(1).toPromise()
-function observableToPromise<T>(observable: Observable<T>): Promise<T> {
-  return new Promise((resolve, reject) => {
-    const subscription = observable.subscribe({
-      next: v => {
-        resolve(v);
-        subscription.unsubscribe();
-      },
-      error: reject,
-      complete: () => {
-        reject(new Error('no value resolved'));
-      },
-    });
-  });
-}
-
-// Duck-type observable detection.
-function isObservable<T>(value: any): value is Observable<T> {
-  return (
-    typeof value === 'object' &&
-    'subscribe' in value &&
-    typeof value.subscribe === 'function'
-  );
-}
-
-function isAsyncIterable(input: unknown): input is AsyncIterable<unknown> {
-  return (
-    typeof input === 'object' &&
-    input !== null &&
-    // Some browsers still don't have Symbol.asyncIterator implemented (iOS Safari)
-    // That means every custom AsyncIterable must be built using a AsyncGeneratorFunction (async function * () {})
-    ((input as any)[Symbol.toStringTag] === 'AsyncGenerator' ||
-      Symbol.asyncIterator in input)
-  );
-}
-
-function asyncIterableToPromise<T>(
-  input: AsyncIterable<T> | AsyncIterableIterator<T>,
-): Promise<T> {
-  return new Promise((resolve, reject) => {
-    // Also support AsyncGenerator on Safari iOS.
-    // As mentioned in the isAsyncIterable function there is no Symbol.asyncIterator available
-    // so every AsyncIterable must be implemented using AsyncGenerator.
-    const iteratorReturn = ('return' in input
-      ? input
-      : input[Symbol.asyncIterator]()
-    ).return?.bind(input);
-    const iteratorNext = ('next' in input
-      ? input
-      : input[Symbol.asyncIterator]()
-    ).next.bind(input);
-
-    iteratorNext()
-      .then(result => {
-        resolve(result.value);
-        // ensure cleanup
-        iteratorReturn?.();
-      })
-      .catch(err => {
-        reject(err);
-      });
-  });
-}
-
-function fetcherReturnToPromise(
-  fetcherResult: FetcherReturnType,
-): Promise<FetcherResult> {
-  return Promise.resolve(fetcherResult).then(fetcherResult => {
-    if (isAsyncIterable(fetcherResult)) {
-      return asyncIterableToPromise(fetcherResult);
-    } else if (isObservable(fetcherResult)) {
-      return observableToPromise(fetcherResult);
-    }
-    return fetcherResult;
-  });
-}
 
 // Determines if the React child is of the same type of the provided React component
 function isChildComponentType<T extends ComponentType>(
